Update README.md
We found this code online which includes characters and allows for different length PIN codes. We can compare this to our code where we will only crack a 4 digit PIN number without characters to show how much more secure a password becomes when more is added to it allowing banks to prevent fraud exponentially greater.